// client/src/lib/api/generated/types/MeOut.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/

import type { MembershipOut } from './MembershipOut'

export type MeOut = {
    /**
     * @description
     * Format: `uuid`
     * @type string
    */
    id: string;
    email: (string | null);
    display_name: (string | null);
    avatar_url: (string | null);
    memberships: MembershipOut[];
};