// client/src/lib/api/generated/zod/meOutSchema.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/

import * as z from 'zod'
import { membershipOutSchema } from './membershipOutSchema'

export const meOutSchema = z.object({
  id: z.uuid(),
  email: z.union([z.string(), z.null()]),
  display_name: z.union([z.string(), z.null()]),
  avatar_url: z.union([z.string(), z.null()]),
  memberships: z.array(membershipOutSchema),
})